requirements.txt
checkerboard==0.2.4
cvxopt==1.3.1
cvxpy==1.3.1
mayavi==4.8.1
motion3d==0.1.1
networkx==3.1
numpy==1.24.3
open3d==0.17.0
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
pandas==2.0.1
PyQt5==5.15.10
PyYAML==6.0
scipy==1.10.1
tabulate==0.9.0
tikzplotlib==0.10.1
transforms3d==0.4.1
vtk==9.2.6
